JSON sub-object logic.
An important part of the full delegation contract will be taking
a user-supplied JSON string of a sub-object, and building up a full
move that has this value placed at a particular path below the
top-level (e.g. g->foobar->VALUE).

This needs proper validation logic to ensure that a maliciously
crafted value (that is invalid JSON) cannot inject a string to
bypass this, e.g. with something like:

  null},"other field":{...

This commit adds a Solidity library with tests for doing this validation
and building up the full JSON string based on a path and sub-object.