Implement move permissions logic.
This adds a base contract MovePermissions, which implements the storage
and handling of our hierarchical permissions for moves.  It has functionality
to check if a particular address is permitted to send a particular move,
and also functions for granting and revoking permissions according to the
proposed rules.

The contract itself does not yet actually implement sending of moves,
but the delegator contract will inherit from MovePermissions to offload
all complexities associated to the permissions.