Create NamePermissions base contract.
This introduces the NamePermissions contract.  It stores a tree of
permissions for each pair of (tokenId, owner), and thus holds the actual
storage of permissions needed for the final delegation contract (which
will inherit from it).  It also provides public methods for interacting
with the permissions, enforcing the right access for operations like
granting or revoking permissions.  It also provides "public view" methods
that can be used to read the entire state, e.g. from a frontend.

Under the hood, all the logic for updating the permissions is done with
the MovePermissions library.  The NamePermissions do not yet link to the
XayaAccounts contract in any way, they just hold the actual permissions
storage under the assumption of a particular token and owner, and allow
updates to them with proper access controls.